fix(landing): server-render landing page instead of auth placeholder
The landing page was wrapped in LandingPageBody, a client component
whose initial render returned a blank aria-busy shell while
useAuth() reported loading=true. On SSR that shell is what Next
serialized, so crawlers and no-JS visitors only ever saw the empty
placeholder — neutralizing the L5.1 marketing content and the L5.2
SEO work that depends on it.

Split the composition: app/page.tsx is now a server component that
renders TopNav, Hero, FeatureTiles, SecondCta, LandingFooter directly
in the initial HTML. LandingAuthRedirect is a new client island that
renders null and handles the authenticated-user redirect to /dashboard
(or /setup) after hydration. LandingPageBody is deleted — no callers.

Verified locally: SSR HTML grew from ~25KB to ~60KB, aria-busy shell
is gone, hero headline + both CTAs + footer copy all present in the
initial response.
